+++ packages/generate-project/src/formatScaffoldToProject.ts
@@ -1,9 +1,9 @@
-import formatFilename from './formatFilename';
-import formatProject from './formatProject';
-import ejsRenderDir from './ejsRenderDir';
 import * as readFiles from 'fs-readdir-recursive';
 import * as fse from 'fs-extra';
 import * as path from 'path';
+import formatFilename from './formatFilename';
+import formatProject from './formatProject';
+import ejsRenderDir from './ejsRenderDir';
 
 export default async function formatScaffoldToProject(projectDir: string, projectName?: string, ejsOptions: any = {}) {
   // format filename