+++ packages/iceworks-cli/src/command/start/index.ts
@@ -1,8 +1,0 @@
-import log from '../../utils/log';
-
-export default async function start() {
-  console.log('');
-  log.warn('', 'Please use Iceworks VS Code plugins: https://ice.work/docs/iceworks/about');
-  console.log('');
-  process.exit(-1);
-}