+++ packages/iceworks-cli/src/command/sync/fusionSDK.ts
@@ -35,7 +35,7 @@ export default class FusionSDK {
   public async getToken() {
     const helpUrl = `${this.fusionHost}/help.html#/dev-create-site`;
     console.log();
-    console.log(`如果这是你第一次使用该功能，或者不知道如何获取Token。\n请查看文档: ${chalk.yellow(helpUrl)}`);
+    console.log(`如果这是你第一次使用该功能，或者不知道如何获取 token。\n请查看文档: ${chalk.yellow(helpUrl)}`);
     console.log();
 
     const { token } = await inquirer.prompt([
@@ -170,7 +170,7 @@ async function requestFusion(options: AxiosRequestConfig, fusionHost: string) {
       err.noAuth = true;
       console.log();
       console.log();
-      console.log(`鉴权失败，请前往 ${fusionHost} 重新获取 token 或 请站点所有者把你添加为站点成员。`);
+      console.log(`鉴权失败，请前往 ${fusionHost} 重新获取 token 或 请站点所有者把你添加为站点成员，完成后重新执行命令。`);
       console.log(`token 文档: ${chalk.yellow(`${fusionHost}/help.html#/dev-create-site`)}`);
       console.log(`添加成员文档: ${chalk.yellow(`${fusionHost}/help.html#/site-user-management`)}`);
       if (err.response.data.success === false) {